+++ src/rpc/methods/sync.rs
@@ -146,7 +146,7 @@ mod tests {
     use crate::libp2p::NetworkMessage;
     use crate::message_pool::{MessagePool, MpoolRpcProvider};
     use crate::networks::ChainConfig;
-    use crate::rpc::eth::filters::EthEventHandler;
+    use crate::rpc::eth::filter::EthEventHandler;
     use crate::rpc::RPCState;
     use crate::shim::address::Address;
     use crate::state_manager::StateManager;
